REF: implement module for shared constructor functions (#27551)
