+++ pandas/core/arrays/__init__.py
@@ -1,4 +1,3 @@
-from .array_ import array  # noqa: F401
 from .base import (  # noqa: F401
     ExtensionArray,
     ExtensionOpsMixin,
